+++ src/Seccenter20240508/Model/RuleListForListWhiteListsOutput.php
@@ -41,7 +41,7 @@ class RuleListForListWhiteListsOutput implements ModelInterface, ArrayAccess
     protected static $swaggerFormats = [
         'match_content' => null,
         'match_key' => null,
-        'match_type' => 'int32'
+        'match_type' => 'int64'
     ];
 
     /**
