+++ src/Seccenter20240508/Model/VirusHitDataListForGetRaspAlarmSummaryInfoOutput.php
@@ -39,8 +39,8 @@ class VirusHitDataListForGetRaspAlarmSummaryInfoOutput implements ModelInterface
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'hit_len' => 'int32',
-        'offset' => 'int32',
+        'hit_len' => 'int64',
+        'offset' => 'int64',
         'rule_id' => null
     ];
